✅ Update: Terminal and arguments - Class 04

+++ dio-lab-java-terminal-argumentos/src/AboutMe.java
@@ -3,7 +3,7 @@ public class AboutMe {
         String nome = args[0];
         String sobrenome = args[1];
         int idade = Integer.valueOf(args[2]);
-        double altura = Double.valueOf(args[3]);
+        int altura = Integer.valueOf(args[3]);
 
         System.out.println("Olá, me chamo " + nome + " " + sobrenome);
         System.out.println("Tenho " + idade + " anos.");
